back-end/requirements.txt
amqp==5.1.1
androguard==3.3.5
ansicon==1.89.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dexofuzzy==1.7.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-picklefield==3.1
django-q==1.3.9
django-sse==0.4.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.3.2
fonttools==4.40.0
fqdn==1.5.1
future==0.18.3
h11==0.14.0
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.3.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
oauthlib==3.2.2
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.6.0
prettytable @ git+https://github.com/jazzband/prettytable@54ccf40d9aadc1c8690773f68ca1c6177858b6f6
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.15.1
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wsproto==1.2.0
